devtools/server/tests/mochitest/test_setupInParentChild.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1181100 - Test DebuggerServerConnection.setupInParent and DebuggerServer.setupInChild
-->
<head>
  <meta charset="utf-8">
  <title>Mozilla Bug</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script type="application/javascript">
"use strict";

const {require} = ChromeUtils.import("resource://devtools/shared/Loader.jsm");
const {DebuggerClient} = require("devtools/shared/client/debugger-client");
const {DebuggerServer} = require("devtools/server/debugger-server");
const {
  connectToFrame,
} = require("devtools/server/connectors/frame-connector");

const Services = require("Services");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({
    "set": [
      // Always log packets when running tests.
      ["devtools.debugger.log", true],
      ["dom.mozBrowserFramesEnabled", true],
    ],
  }, runTests);
};

function runTests() {
  // Create a minimal iframe with a message manager
  const iframe = document.createElement("iframe");
  iframe.mozbrowser = true;
  document.body.appendChild(iframe);

  // Instantiate a minimal server
  DebuggerServer.init();
  if (!DebuggerServer.createRootActor) {
    DebuggerServer.registerAllActors();
  }

  // Fake a connection to an iframe
  const transport = DebuggerServer.connectPipe();
  const conn = transport._serverConnection;
  const client = new DebuggerClient(transport);

  // Wait for a response from setupInChild
  const onChild = msg => {
    Services.ppmm.removeMessageListener("test:setupChild", onChild);
    const args = msg.json;

    is(args[0], 1, "Got first numeric argument");
    is(args[1], "two", "Got second string argument");
    is(args[2].three, true, "Got last JSON argument");

    // Ask the child to call setupInParent
    DebuggerServer.setupInChild({
      module: "chrome://mochitests/content/chrome/devtools/server/tests/mochitest/setup-in-child.js",
      setupChild: "callParent",
    });
  };
  Services.ppmm.addMessageListener("test:setupChild", onChild);

  // Wait also for a reponse from setupInParent called from setup-in-child.js
  const onParent = (_, topic, args) => {
    Services.obs.removeObserver(onParent, "test:setupParent");
    args = JSON.parse(args);

    is(args[0], true, "Got `mm` argument, a message manager");
    ok(args[1].match(/server\d+.conn\d+.child\d+/), "Got `prefix` argument");

    cleanup();
  };
  Services.obs.addObserver(onParent, "test:setupParent");

  // Instanciate e10s machinery and call setupInChild
  connectToFrame(conn, iframe).then(actor => {
    DebuggerServer.setupInChild({
      module: "chrome://mochitests/content/chrome/devtools/server/tests/mochitest/setup-in-child.js",
      setupChild: "setupChild",
      args: [1, "two", {three: true}],
    });
  });

  function cleanup() {
    client.close().then(function() {
      DebuggerServer.destroy();
      iframe.remove();
      SimpleTest.finish();
    });
  }
}
</script>
</pre>
</body>
</html>
